hf_space/requirements.txt
gradio
torch
torchvision
transformers
pillow
scikit-image
